products.js, handlers/controller/routes
Unificado el estilo en el que se harán los controllers, handlers y routes

El archivo products.js fue movido a la carpeta "data"

Emprolijado general
