+++ README.md
@@ -1,5 +1,7 @@
 PROYECTO FINAL SOY HENRY 2023
+
 COHORTE 37B – GRUPO 12
+
 E-COMMERCE DE VENTA DE PRODUCTOS DE COMPUTACIÓN
 
 Les presentamos nuestro proyecto final: 1337 HARDWARE – Elite Gaming Store. Se tratará de la plataforma e-commerce de un local comercial de venta de productos relacionados a la computación. Si bien el comercio tendrá como especialidad el hardware para gamers, también ofrecerá productos informáticos en general.
