--- server/src/routes/productsRouter.js
@@ -1,11 +1,0 @@
-const { Router } = require("express");
-const { getProducts, getProductByIdHandler } = require('../handlers/productsHandler');
-const postCreateProduct = require('../functions/postCreateProduct');
-
-const productsRouter = Router();
-
-productsRouter.get("/", getProducts);
-productsRouter.get("/:productId", getProductByIdHandler);
-productsRouter.post("/", postCreateProduct);
-
-module.exports = productsRouter;
